# docs/openapi.json
{
  "components": {
    "schemas": {
      "AdvanceRequest": {
        "additionalProperties": false,
        "properties": {
          "approval_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Token"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 8,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Stage"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "AdvanceRequest",
        "type": "object"
      },
      "ApprovalDecision": {
        "enum": [
          "approved",
          "denied"
        ],
        "title": "ApprovalDecision",
        "type": "string"
      },
      "ApprovalDecisionRequest": {
        "additionalProperties": false,
        "properties": {
          "approver": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Approver",
            "type": "string"
          },
          "decision": {
            "$ref": "#/components/schemas/ApprovalDecision"
          },
          "expected_digest": {
            "title": "Expected Digest",
            "type": "string"
          }
        },
        "required": [
          "decision",
          "approver",
          "expected_digest"
        ],
        "title": "ApprovalDecisionRequest",
        "type": "object"
      },
      "AutorunRequest": {
        "additionalProperties": false,
        "properties": {
          "approval_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Token"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 8,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          }
        },
        "title": "AutorunRequest",
        "type": "object"
      },
      "DemoResetRequest": {
        "additionalProperties": false,
        "properties": {
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 8,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "scenario": {
            "default": "happy_path",
            "enum": [
              "happy_path",
              "insufficient_evidence"
            ],
            "title": "Scenario",
            "type": "string"
          }
        },
        "title": "DemoResetRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Stage": {
        "enum": [
          "INTAKE",
          "CONTEXT",
          "PLAN",
          "PLAN_REVIEW",
          "APPROVAL",
          "EXECUTE",
          "OBSERVE",
          "EVALUATE",
          "VERIFY",
          "DECIDE",
          "ARCHIVE",
          "MEMORY_SKILL",
          "COMPLETED"
        ],
        "title": "Stage",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Evidence-gated, deterministic control plane for multi-agent embodied-AI research. The bundled EgoLite run is explicitly synthetic.",
    "title": "EgoAgentOS ResearchOps API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/approvals/{approval_id}/decision": {
      "post": {
        "operationId": "approval_decision_api_v1_approvals__approval_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approval Decision Api V1 Approvals  Approval Id  Decision Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approval Decision",
        "tags": [
          "approval"
        ]
      }
    },
    "/api/v1/dashboard": {
      "get": {
        "operationId": "dashboard_api_v1_dashboard_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Dashboard Api V1 Dashboard Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Dashboard",
        "tags": [
          "research"
        ]
      }
    },
    "/api/v1/demo/reset": {
      "post": {
        "operationId": "reset_demo_api_v1_demo_reset_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DemoResetRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Reset Demo Api V1 Demo Reset Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Demo",
        "tags": [
          "demo"
        ]
      }
    },
    "/api/v1/health": {
      "get": {
        "operationId": "health_api_v1_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Health Api V1 Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "system"
        ]
      }
    },
    "/api/v1/integrations": {
      "get": {
        "operationId": "integrations_api_v1_integrations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Integrations Api V1 Integrations Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Integrations",
        "tags": [
          "system"
        ]
      }
    },
    "/api/v1/tasks": {
      "get": {
        "operationId": "tasks_api_v1_tasks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Tasks Api V1 Tasks Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Tasks",
        "tags": [
          "research"
        ]
      }
    },
    "/api/v1/tasks/{task_id}": {
      "get": {
        "operationId": "task_api_v1_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Task Api V1 Tasks  Task Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Task",
        "tags": [
          "research"
        ]
      }
    },
    "/api/v1/tasks/{task_id}/advance": {
      "post": {
        "operationId": "advance_task_api_v1_tasks__task_id__advance_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AdvanceRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Advance Task Api V1 Tasks  Task Id  Advance Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Advance Task",
        "tags": [
          "research"
        ]
      }
    },
    "/api/v1/tasks/{task_id}/autorun": {
      "post": {
        "operationId": "autorun_task_api_v1_tasks__task_id__autorun_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AutorunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Autorun Task Api V1 Tasks  Task Id  Autorun Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Autorun Task",
        "tags": [
          "research"
        ]
      }
    },
    "/api/v1/tasks/{task_id}/events": {
      "get": {
        "operationId": "task_events_api_v1_tasks__task_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_sequence",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Sequence",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Task Events Api V1 Tasks  Task Id  Events Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Task Events",
        "tags": [
          "audit"
        ]
      }
    }
  }
}